fix(debug mode): wrong reason for why RBR reports are visible in LHN
